refactor: pick out modal to separate file
