--- src/assets/scripts/data/pages.ts
@@ -1,17 +1,0 @@
-import { PageIds } from "../types/enums";
-import { stringObject } from "../types/Interfaces";
-
-export const Buttons: stringObject[] = [
-  {
-    id: PageIds.MainPage,
-    text: 'Main Page',
-  },
-  {
-    id: PageIds.BasketPage,
-    text: 'Basket Page',
-  },
-  {
-    id: PageIds.BookPage,
-    text: 'Book Page',
-  },
-];
