+++ src/assets/styles/_variables.scss
@@ -9,7 +9,7 @@ html {
   font-family: 'Cinzel', serif;
   color: #041e41;
   font-weight: 400;
-  font-size: 25px;
+  font-size: 24px;
 }
 
 body{
